+++ python/GafferSceneUI/__init__.py
@@ -48,6 +48,7 @@ from .UVInspector import UVInspector
 from .FilterPlugValueWidget import FilterPlugValueWidget
 from .ScenePathPlugValueWidget import ScenePathPlugValueWidget
 from .LightEditor import LightEditor
+from .SetEditor import SetEditor
 from . import SceneHistoryUI
 from . import EditScopeUI